requirements.txt
Django==2.2.19
djangorestframework==3.8.2
django-cors-headers==2.4.0